refactor(refs T30969): move DpButtonRow to components root
