+++ src/components/core/DpEditor/DpLinkModal.vue
@@ -46,7 +46,7 @@
 </template>
 
 <script>
-import DpButtonRow from '../DpButtonRow'
+import DpButtonRow from '../../DpButtonRow/DpButtonRow'
 import DpCheckbox from '../../DpCheckbox/DpCheckbox'
 import DpInput from '../../DpInput/DpInput'
 import DpModal from '../DpModal'
