+++ src/components/index.js
@@ -1,4 +1,5 @@
 import DpButton from './DpButton/DpButton'
+import DpButtonRow from './DpButtonRow/DpButtonRow'
 import DpCheckbox from './DpCheckbox/DpCheckbox'
 import DpDetails from './DpDetails/DpDetails'
 import DpIcon from './DpIcon/DpIcon'
@@ -12,7 +13,6 @@ import {
   DpAutocomplete,
   DpBulkEditHeader,
   DpButtonIcon,
-  DpButtonRow,
   DpCard,
   DpChangeStateAtDate,
   DpCheckboxGroup,
@@ -69,6 +69,7 @@ import {
 
 export {
   DpButton,
+  DpButtonRow,
   DpDetails,
   DpIcon,
   DpInput,
@@ -80,7 +81,6 @@ export {
   DpAutocomplete,
   DpBulkEditHeader,
   DpButtonIcon,
-  DpButtonRow,
   DpCard,
   DpChangeStateAtDate,
   DpCheckbox,
